+++ karavan-designer/src/designer/api/CamelApi.tsx
@@ -1165,3 +1165,4 @@ export class CamelApi {
         return undefined;
     }
 }
+